+++ subterra-optim/src/main/java/io/toterra/subterra/optim/client/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Client optimizations: particle/entity culling, idle throttling.
+ */
+package io.toterra.subterra.optim.client;