+++ subterra-optim/src/main/java/io/toterra/subterra/optim/entity/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Entity optimizations: entity & AI, XP aggregation.
+ */
+package io.toterra.subterra.optim.entity;