+++ subterra-optim/src/main/java/io/toterra/subterra/optim/logic/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Logic optimizations: redstone, recipe/container caching, micro-optimizations.
+ */
+package io.toterra.subterra.optim.logic;