+++ subterra-optim/src/main/java/io/toterra/subterra/optim/worldgen/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * World-gen optimizations: terrain noise, density-function evaluation caching,
+ * lighting generation.
+ */
+package io.toterra.subterra.optim.worldgen;